refactor: 분석 화면 렌더링 경계 정리
