+++ apps/fe/src/app/(main)/analysis/[id]/page.tsx
@@ -1,17 +1,14 @@
 'use client';
 
 import { use, useState } from 'react';
-import { ChatBubble, ToastPortal } from '@/components';
+import { ToastPortal } from '@/components';
 import { useAuthSession } from '@/providers/AuthProvider';
 import PromptInput from '../_components/PromptInput';
-import AnalysisResult from './_components/AnalysisResult';
-import AnalyzingIndicator from './_components/AnalyzingIndicator';
+import AnalysisChatMessageList from './_components/AnalysisChatMessageList';
 import DataTablePreview from './_components/DataTablePreview';
 import LoadingDataPreview from './_components/LoadingDataPreview';
-import QuestionAnalysisResult from './_components/QuestionAnalysisResult';
 import ResizableSplit from './_components/ResizableSplit';
-import VerificationResult from './_components/VerificationResult';
-import { useAnalysisChat, type ChatMessage } from './_hooks/useAnalysisChat';
+import { useAnalysisChat } from './_hooks/useAnalysisChat';
 
 type PageParams = { id: string };
 
@@ -28,105 +25,6 @@ export default function AnalysisChatPage({
     hasAccessToken: isAuthenticated,
     routeConversationId: id,
   });
-
-  const renderSummaryMessage = () => {
-    if (!chat.summary) return null;
-
-    const hasWarnings = chat.summaryWarnings.length > 0;
-
-    return (
-      <div key="summary" className="flex w-full justify-start">
-        <div className="w-full max-w-[80%]">
-          <AnalysisResult
-            summary={chat.summary}
-            warningActions={
-              hasWarnings
-                ? {
-                    disabled: chat.summaryActionDisabled,
-                    onEdit: () => chat.startInitialQuestion('edit'),
-                    onContinue: () => chat.startInitialQuestion(),
-                  }
-                : undefined
-            }
-          />
-        </div>
-      </div>
-    );
-  };
-
-  const renderErrorMessage = () => {
-    if (!chat.summaryErrorMessage) return null;
-
-    return (
-      <ChatBubble key="summary-error" role="bot">
-        <p className="text-error-500">{chat.summaryErrorMessage}</p>
-      </ChatBubble>
-    );
-  };
-
-  const renderMessage = (message: ChatMessage) => {
-    if (message.role === 'user') {
-      return (
-        <ChatBubble
-          key={message.id}
-          role="user"
-          file={
-            message.fileName
-              ? { name: message.fileName, status: 'uploaded' }
-              : undefined
-          }
-        >
-          {message.content}
-        </ChatBubble>
-      );
-    }
-
-    if (message.kind === 'notice') {
-      return (
-        <ChatBubble key={message.id} role="bot">
-          <p
-            className={
-              message.tone === 'error' ? 'text-error-500' : 'text-gray-900'
-            }
-          >
-            {message.content}
-          </p>
-        </ChatBubble>
-      );
-    }
-
-    if (message.kind === 'verification') {
-      return (
-        <div key={message.id} className="flex w-full justify-start">
-          <div className="w-full max-w-[80%]">
-            <VerificationResult result={message.result} />
-          </div>
-        </div>
-      );
-    }
-
-    return (
-      <div key={message.id} className="flex w-full justify-start">
-        <div className="w-full max-w-[80%]">
-          <QuestionAnalysisResult
-            criteria={message.criteria}
-            initialMode={message.initialMode}
-            baseDateColumnOptions={
-              chat.summary?.dateFieldOptions.length
-                ? chat.summary.dateFieldOptions
-                : chat.summaryColumnOptions
-            }
-            groupByOptions={chat.summaryColumnOptions}
-            sortByOptions={chat.summarySortOptions}
-            isSubmitting={chat.criteriaSubmitting}
-            onContinue={(values) =>
-              chat.handleConfirmCriteria(message.criteria.messageId, values)
-            }
-          />
-        </div>
-      </div>
-    );
-  };
 
   return (
     <ResizableSplit
@@ -147,17 +45,7 @@ export default function AnalysisChatPage({
       right={
         <div className="flex h-full min-h-0 flex-col">
           <div className="scrollbar-hide flex min-h-0 flex-1 flex-col gap-24 overflow-y-auto px-24 pt-32 pb-24">
-            {chat.initialMessage ? renderMessage(chat.initialMessage) : null}
-            {renderSummaryMessage()}
-            {renderErrorMessage()}
-            {chat.restMessages.map(renderMessage)}
-            {chat.shouldShowAnalyzing && (
-              <div className="flex w-full justify-start">
-                <div className="w-full max-w-[80%]">
-                  <AnalyzingIndicator message={chat.analyzingMessage} />
-                </div>
-              </div>
-            )}
+            <AnalysisChatMessageList chat={chat} />
           </div>
 
           <div className="shrink-0 px-24 pt-12 pb-24">
